remove Image widget - functionality was added to frame
Took 20 minutes
